Rename QueryHooks -> GraphQLHooks.
